requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0
psutil>=5.9.0

# Quantization
auto-gptq>=0.7.0
optimum>=1.14.0

# PiSSA
# Install from: https://github.com/hiyouga/PiSSA

# GaLore
galore-torch>=1.0.0

# SpinQuant
# Install from: https://github.com/ModelTC/SpinQuant

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
tensorboard>=2.14.0
wandb>=0.15.0

# Data processing
datasets>=2.14.0

# Model specific
qwen-vl-utils>=0.0.1

# Benchmarking dependencies
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Code execution (for LiveCodeBench and SWE-Bench)
# Optional: Install if you need code execution capabilities
# docker>=6.1.0
